+++ frontend/app/organizations/[org_slug]/page.tsx
@@ -1,6 +1,7 @@
 import { Spacer } from "@/components/nextui-use-client";
 import TournamentsTable from "@/components/tournaments-table";
-import { getOrganization, getOrganizations, getOrganizationTournaments } from "@/app/data/actions";
+import { getOrganization, getOrganizations } from "@/app/data/organizations/actions";
+import { getOrganizationTournaments } from "@/app/data/organizations/tournaments/actions";
 
 import OrgDetailCard from "@/app/organizations/[org_slug]/org-detail-card";
