+++ frontend/lib/api/client.ts
@@ -4,6 +4,14 @@ import { getVercelOidcToken } from "@vercel/functions/oidc";
 import { auth } from "@clerk/nextjs/server";
 import { paths } from "@/lib/api/openapi-v1";
 
+export const DEFAULT_CACHE_TIMEOUT: number = 300;
+
+export function defaultConfig(tag: string, revalidate?: number) {
+  return {
+    next: { tags: [tag], revalidate: revalidate ?? DEFAULT_CACHE_TIMEOUT },
+  };
+}
+
 export function getBaseUrl() {
   if ([env.NODE_ENV, env.VERCEL_ENV].includes("production")) {
     return `${env.PROD_API_BASE_URL}`;
